Move GetUserAgentHeader to internal packages
